refactor: organise dynamic_input_amount JS logic into folders
