+++ nodes/web/helpers/dynamic_input_amount/main.js
@@ -1,0 +1,3 @@
+import { classNames, extNames, handleNumSlots, updateTypesBasedOnConnection, widgetLogic } from "./utils.js";
+
+export { classNames, extNames, handleNumSlots, updateTypesBasedOnConnection, widgetLogic };
